Bug 1506364 - Implement the prefers-contrast media-query. r=morgan,emilio
Differential Revision: https://phabricator.services.mozilla.com/D79553

+++ layout/style/ServoStyleSet.h
@@ -26,7 +26,7 @@
 #include "nsIMemoryReporter.h"
 
 namespace mozilla {
-enum class MediaFeatureChangeReason : uint8_t;
+enum class MediaFeatureChangeReason : uint16_t;
 namespace css {
 class Rule;
 }  // namespace css
